backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.3.0
cloudinary==1.36.0
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
